axi_dmac: Add support for DMA Scatter-Gather
This commit introduces a different interface to submit transfers, using
DMA descriptors.

The structure of the DMA descriptor is as follows:

struct dma_desc {
    u32 flags,
    u32 id,
    u64 dest_addr,
    u64 src_addr,
    u64 next_sg_addr,
    u32 y_len,
    u32 x_len,
    u32 src_stride,
    u32 dst_stride,
};

The 'flags' field currently offers two control bits:
- bit 0: if set, the transfer will complete after this last descriptor
  is processed, and the DMA core will go back to idle state.
  If clear, the next DMA descriptor (which address is contained
  in 'next_sg_addr') will be loaded.
- bit 1: if set, an end-of-transfer interrupt will be raised after the
  transfer corresponding to this descriptor completed.

The 'id' field corresponds to an identifier of the descriptor.
When a descriptor has been loaded, register 0x454 is updated with the
value of this field.

The 'dest_addr' and 'src_addr' contain the destination and source
addresses to use for the transfer, respectively.

The 'x_len' field contains the number of bytes to transfer,
minus one.

The 'y_len', 'src_stride' and 'dst_stride' fields are only useful for
2D transfers, and should be set to zero if 2D transfers are not
required.

To start a transfer, the address of the first DMA descriptor must be
written to register 0x47c and the HWDESC bit of CONTROL register must
be set.

Signed-off-by: Ionut Podgoreanu <ionut.podgoreanu@analog.com>